requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
google-generativeai==0.3.2
python-dotenv==1.0.1
aiohttp==3.9.3
pytest==8.0.0
pytest-asyncio==0.23.5
httpx==0.26.0
python-multipart==0.0.9
pytest-cov==4.1.0
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1 